Add **kwargs to allow for parglare options
Now, all the normal parglare options can be passed in directly to the
JSONParser methods and Grammar/Parser will use them accordingly.

Also added the functionality to specify whether you want the output to
be "cleaned" (default is True), meaning that the extra grammar actions
are executed, returning the JSON as a Python dict with no extra lists or
quotation marks.

If cleaned=False, no grammar actions are applied, leaving the output as
strings or lists of strings.