Add `build_tree` functionality
The build_tree kwarg must be specified during instantiation, NOT passed
directly to one of the methods, as Grammar doesn't take that kwarg.

If `build_Tree` is False, an info message will be printed to the
terminal to let the user know how to view the parse tree if desired.